# backend/app/docs/swagger.yaml
definitions:
...
  app_internal_handler_item.CreateItemReq:
    properties:
      content:
        example: 这是一个项目
        maxLength: 1000
        minLength: 3
        type: string
      status:
        allOf:
        - $ref: '#/definitions/backend_app_types_meta.ItemStatus'
        enum:
        - normal
        - done
        - marked
        example: normal
      tags:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        maxItems: 10
        minItems: 1
        type: array
    required:
    - content
    type: object
  app_internal_handler_item.GetDailyItemCountResp:
    properties:
      daily_item_counts:
        items:
          $ref: '#/definitions/backend_app_types_dto.DailyItemCountDTO'
        type: array
    type: object
  app_internal_handler_item.GetItemListResp:
    properties:
      items:
        items:
          $ref: '#/definitions/backend_app_types_dto.ItemDTO'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  app_internal_handler_item.UpdateItemReq:
    properties:
      content:
        example: 这是一个项目
        maxLength: 1000
        minLength: 3
        type: string
      status:
        allOf:
        - $ref: '#/definitions/backend_app_types_meta.ItemStatus'
        enum:
        - normal
        - done
        - marked
        example: normal
      tags:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        maxItems: 10
        minItems: 1
        type: array
    type: object
...
  backend_app_types_dto.DailyItemCountDTO:
    properties:
      count:
        type: integer
      date:
        type: string
    type: object
  backend_app_types_dto.ItemDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      item_id:
        type: integer
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/backend_app_types_dto.TagDTO'
        type: array
      updated_at:
        type: string
    type: object
  backend_app_types_dto.TagDTO:
    properties:
      color:
        type: string
      icon:
        type: string
      tag_id:
        type: integer
      tag_name:
        type: string
      tag_value:
        type: string
    type: object
  backend_app_types_meta.ItemStatus:
    enum:
    - normal
    - done
    - marked
    type: string
    x-enum-varnames:
    - ItemStatusNormal
    - ItemStatusDone
    - ItemStatusMarked
  backend_utils_handle.Response:
    properties:
      code:
        description: 响应码，0 表示成功
        example: 0
        type: integer
      data:
        description: 响应数据（可选）
      message:
        description: 响应消息（可选）
        example: 操作成功
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于 Go 和 Gin 框架的 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /api/item:
    post:
      consumes:
      - application/json
      description: 创建一个新项目
      parameters:
      - description: 创建项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_internal_handler_item.CreateItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.ItemDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 创建项目
      tags:
      - 项目管理
  /api/item/{item_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定项目
      parameters:
      - description: 项目ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 删除项目
      tags:
      - 项目管理
    get:
      consumes:
      - application/json
      description: 获取指定项目的详细信息
      parameters:
      - description: 项目ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.ItemDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 获取项目
      tags:
      - 项目管理
    put:
      consumes:
      - application/json
      description: 更新指定项目的信息
      parameters:
      - description: 项目ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: 更新项目请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_internal_handler_item.UpdateItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.ItemDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 更新项目
      tags:
      - 项目管理
  /api/item/daily-count:
    get:
      consumes:
      - application/json
      description: 获取每日项目数量
      parameters:
      - description: 开始日期
        in: query
        name: date_start
        required: true
        type: string
      - description: 结束日期
        in: query
        name: date_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_internal_handler_item.GetDailyItemCountResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 获取每日项目数量
      tags:
      - 项目管理
  /api/item/list:
    get:
      consumes:
      - application/json
      description: 获取项目列表，支持分页和筛选
      parameters:
      - description: 开始日期
        in: query
        name: date_start
        type: string
      - description: 结束日期
        in: query
        name: date_end
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_internal_handler_item.GetItemListResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 获取项目列表
      tags:
      - 项目管理
...
schemes:
- http
- https
swagger: "2.0"